pyproject.toml
[project]
name = "openai-embedding"
version = "0.1.0"
description = ""
authors = [{ name = "sunjbs", email = "sunjb98@gmail.com" }]
readme = "README.md"
requires-python = ">=3.12, <3.15"
dependencies = [
    "openai (>=1.102.0,<2.0.0)",
    "pypdf (>=6.0.0,<7.0.0)",
    "faiss-cpu (>=1.12.0,<2.0.0)",
    "numpy (>=2.3.2,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
]

[tool.poetry]
packages = [{ include = "openai_embed", from = "src" }]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
